feat: add rarity for npc pickpocketing
