+++ src/main/resources/thieving.json
@@ -1,0 +1,1 @@
+{"Guard":[{"i":2809,"d":128,"q":1},{"i":22879,"d":896,"q":2}],"Gnome":[{"i":2809,"d":150,"q":1}],"H.A.M. Member":[{"i":4298,"d":100,"q":1},{"i":4300,"d":100,"q":1},{"i":4302,"d":100,"q":1},{"i":4304,"d":100,"q":1},{"i":4306,"d":100,"q":1},{"i":4308,"d":100,"q":1},{"i":4310,"d":100,"q":1}],"Hero":[{"i":12157,"d":1400,"q":1}],"Master Farmer":[{"i":5295,"d":302,"q":1},{"i":5296,"d":443,"q":1},{"i":5298,"d":947,"q":1},{"i":5299,"d":1389,"q":1},{"i":5300,"d":2083,"q":1},{"i":5301,"d":2976,"q":1},{"i":5303,"d":6944,"q":1},{"i":5304,"d":10417,"q":1},{"i":22879,"d":260,"q":1}],"Paladin":[{"i":3560,"d":1000,"q":1}],"TzHaar-Hur":[{"i":1617,"d":195,"q":1}],"Wealthy citizen":[{"i":2711,"d":85,"q":1}]}
